+++ external_plugins/power_measuring_plugin/sampleout.json
@@ -1,0 +1,56 @@
+// this is a sample output from jtop
+{
+    "time": datetime.datetime(2023,
+    10,
+    19,
+    14,
+    30,
+    30,
+    763607),
+    "uptime": datetime.timedelta(days=1, seconds=3123, microseconds=700000),
+    "CPU1": 7,
+    "CPU2": 12,
+    "CPU3": 5,
+    "CPU4": 4,
+    "CPU5": 4,
+    "CPU6": 1,
+    "CPU7": 9,
+    "CPU8": 2,
+    "CPU9": 11,
+    "CPU10": 5,
+    "CPU11": 16,
+    "CPU12": 6,
+    "RAM": 0.14743634686947704,
+    "SWAP": 0.0,
+    "EMC": 0,
+    "GPU": 0.0,
+    "APE": "OFF",
+    "DLA0_CORE": "OFF",
+    "DLA0_FALCON": "OFF",
+    "DLA1_CORE": "OFF",
+    "DLA1_FALCON": "OFF",
+    "NVDEC": "OFF",
+    "NVENC": "OFF",
+    "NVJPG": "OFF",
+    "NVJPG1": "OFF",
+    "PVA0_CPU_AXI": "OFF",
+    "PVA0_VPS": "OFF",
+    "SE": 473600,
+    "VIC": "OFF",
+    "Fan pwmfan0": 18.823529411764707,
+    "Temp CPU": 47.562,
+    "Temp CV0": -256.0,
+    "Temp CV1": -256.0,
+    "Temp CV2": -256.0,
+    "Temp GPU": 42.375,
+    "Temp SOC0": 44.187,
+    "Temp SOC1": 43.562,
+    "Temp SOC2": 43.656,
+    "Temp Tboard": 37.0,
+    "Temp Tdiode": 37.75,
+    "Temp tj": 47.812,
+    "Power tj": 4247,
+    "Power TOT": 9637,
+    "jetson_clocks": "OFF",
+    "nvp model": "MAXN"
+}
