+++ releases/0.3.3_power_test/config/image_gen_config.json
@@ -1,0 +1,6 @@
+{
+    "path": "/example_images",
+    "timestamp": "600",
+    "callingFunction": "simpleNext",
+    "burstQuantity": "5"
+}
