+++ releases/0.3.3_power_test/docker-compose.yml
@@ -1,0 +1,72 @@
+version: "3.0"
+
+networks:
+  cameratraps:
+    driver: bridge
+
+services:
+  # the name `engine` is important here; in general, the sevice name is addressable by other conatiners on the same
+  # docker network. The default "hostname" used by the python plugin library (pyevents) is "engine"
+  engine:
+    container_name: engine
+    image: tapis/camera_traps_engine:0.3.3
+    networks:
+      - cameratraps
+    environment:
+      - TRAPS_CONFIG_FILE=/traps.toml
+    volumes:
+      # mount the traps.toml in the current working directory.
+      - ./config/traps.toml:/traps.toml:ro
+      # mount the image output directory from the host to the directory specified in traps.toml
+      # Docker compose hijacks $HOME so we use a workaround.  If the source directory doesn't
+      # exist it will be created with root ownership.
+      - ./images_output_dir:/root/camera-traps/images
+      # mount the log4rs configuration file over the baked into the image.  Comment out
+      # this mount if you want to use the image's default logging configuration.
+      - ../../resources/log4rs.yml:/resources/log4rs.yml
+
+  powerMonitorPlugin:
+    # USAGE Example:
+    # In your plugins, import the necessary:
+    # E.g. Image_scoring_plugin.py:
+    #
+    #   from ctevents.ctevents import send_power_measure_fb_event
+    #   import os
+    #
+    # When trying to measure power:
+    #     my_pids = [os.getpid()] # any pids you want to measure
+    #     monitor_type = [0] # 0 for CPU & GPU, 1 for CPU, 2 for GPU  (TODO: 3 for DRAM)
+    #     monitor_duration = 10 # seconds
+    #     send_power_measure_fb_event(socket, my_pids, monitor_type, monitor_duration)
+    #
+    container_name: power_monitor
+    image: tapis/power_measuring_plugin_py:0.3.3
+    networks:
+      - cameratraps
+    pid: host
+    # Uncomment the following line to use GPUs with the Power Monitor pluging. Note that
+    # the NVIDIA drivers must be installed, the NVIDIA container toolkit installed, and 
+    # Docker must be configured to use it. See the README for details.
+    # UNCOMMENT --------------    
+    # deploy:
+    #   resources:
+    #     reservations:
+    #       devices:
+    #         - driver: nvidia
+    #           count: all # Use 'all' or specify the number of GPUs you want to use
+    #           capabilities: [ gpu ]
+    # ------------------------
+    depends_on:
+      - engine
+
+    environment:
+      - POWER_MEASURING_PLUGIN_PORT=6010
+      # Note: this LOG_PATH is the path in the conatiner, so must matcht the RHS of the mount below.
+      - TRAPS_POWER_LOG_PATH=/logs
+      - TRAPS_TEST_POWER_FUNCTION=1
+      # 1 for debug; Set to 0 when actually use it to measure other plugin's power
+    volumes:
+      # mount the traps.toml in the current working directory.
+      - ./config/traps.toml:/traps.toml:ro
+      - ./power_output_dir:/logs # Path for CPU and GPU power logs
+      - /run/jtop.sock:/run/jtop.sock # TODO: Required for JTOP to work
